// frontend/menza-web/src/api/models/meal.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Meal {
  id?: number;
  name?: string;
  description?: string;
  imageUrl?: string;
  weight?: number;
  kcal?: number;
  fats?: number;
  proteins?: number;
  carbs?: number;
  category?: string;
}